requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.5.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0 
django-filter==23.2
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
python-magic==0.4.27
Faker==19.6.2
